+++ src/components/Liquidity.tsx
@@ -1,9 +1,9 @@
-import { Token, Pair, getPairByLauncherId, getLiquidityQuote, ActionType } from '../../api';
+import { Token, Pair, getPairByLauncherId, getLiquidityQuote, ActionType } from '../api';
 import { UNKNWN, XCH, getLiquidityToken } from '@/shared_tokens';
-import GenerateOfferButton from '../atomic/GenerateOfferButton';
-import BooleanSwitch from '../atomic/BooleanSwitch';
-import TokenSelector from '../atomic/TokenSelector';
-import { GenerateOfferData } from '../TabContainer';
+import GenerateOfferButton from './atomic/GenerateOfferButton';
+import BooleanSwitch from './atomic/BooleanSwitch';
+import TokenSelector from './atomic/TokenSelector';
+import { GenerateOfferData } from './TabContainer';
 import LiquidityInput from './LiquidityInput';
 import { useState, useEffect } from 'react';
